+++ packages/cached/test/cached_getter_generation_test.dart
@@ -18,7 +18,7 @@ Future<void> main() async {
     'AsyncSyncWrite',
     'SyncSyncWrite',
     'PersistentCachedGetter',
-    'LazyPersistentCachedGetter',
+    'DirectPersistentCachedGetter',
     'InitOnCallPersistentCachedGetter',
   };