+++ packages/cached/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.7.0
 * Fix not working peristent storage with the getters
-* Add `@LazyPersistentCached()` annotation
+* Add `@DirectPersistentCached()` annotation
 * Add `@PersistentCached()` annotation
 * Add `@PersistentCached(initOnCall: )` parameter - initialize cache from external storage only after the first method call
 * Mark `@Cached(persistentStorage: )` as @Deprecated